Move simple components into components/index
