--- client/src/components/centered-circular-progress.js
@@ -1,5 +1,0 @@
-import CircularProgress from '@material-ui/core/CircularProgress';
-import styled from 'react-emotion';
-
-const CenteredCircularProgress = styled(CircularProgress)({margin: 'auto'});
-export default CenteredCircularProgress;
