+++ client/src/components/login-form.js
@@ -2,9 +2,9 @@ import Button from '@material-ui/core/Button';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
-import FormField from './form-field';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
+import {FormField} from '.';
 import {connect} from 'react-redux';
 import {logIn} from '../actions/user';
