--- client/src/components/styled-dialog-content.js
@@ -1,8 +1,0 @@
-import DialogContent from '@material-ui/core/DialogContent';
-import styled from 'react-emotion';
-
-const StyledDialogContent = styled(DialogContent)({
-  overflowY: 'visible'
-});
-
-export default StyledDialogContent;
