+++ client/src/pages/skater/skater-content.js
@@ -5,7 +5,6 @@ import Helmet from 'react-helmet';
 import PropTypes from 'prop-types';
 import React, {Component, Fragment} from 'react';
 import SkaterForm from './skater-form';
-import StyledDialogContent from '../../components/styled-dialog-content';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
@@ -16,6 +15,7 @@ import compose from 'recompose/compose';
 import differenceInYears from 'date-fns/differenceInYears';
 import round from 'lodash/round';
 import withProps from 'recompose/withProps';
+import {StyledDialogContent} from '../../components';
 import {connect} from 'react-redux';
 import {countries} from 'countries-list';
 import {getAggregates, getLetters, getRoundName} from '../../util/game';
