+++ client/src/pages/skater/skater-form.js
@@ -3,10 +3,6 @@ import DialogActions from '@material-ui/core/DialogActions';
 import DialogButton from '../../components/dialog-button';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
-import FormField, {
-  FormFieldControl,
-  formFieldProps
-} from '../../components/form-field';
 import InputLabel from '@material-ui/core/InputLabel';
 import MenuItem from '@material-ui/core/MenuItem';
 import PropTypes from 'prop-types';
@@ -16,6 +12,7 @@ import gql from 'graphql-tag';
 import sortBy from 'lodash/sortBy';
 import upperFirst from 'lodash/upperFirst';
 import {DatePicker} from 'material-ui-pickers';
+import {FormField, FormFieldControl, formFieldProps} from '../../components';
 import {Mutation} from 'react-apollo';
 import {STANCES, STANCE_REGULAR} from '../../../../api/common';
 import {countries} from 'countries-list';
